chore: Install missing packages if required
This commit updates the timeshift_setup.sh script to check for the presence of required packages (timeshift, grub-btrfs, snapd) before proceeding with the installation. If any of the packages are missing, the script will install them using the install_packages.sh script. This ensures that all necessary dependencies are in place for the script to run successfully.

Recent user commits:
- chore: Add snapd package to yay installation list
- chore: Update variable name for YAML file in install_packages.sh
- bugfixing
- chore: Refactor config sourcing to prevent duplicate sourcing and improve efficiency
- set -x
- config source reworked