chore: Update setup.sh to remove unnecessary packages from installation
This commit removes the unnecessary packages from the `setup_install_packages` variable in the `config.sh` file. Previously, the variable included both the `packages_core` and `packages_desktop` sets, but now it only includes the `packages_core` set. This change improves the efficiency of the installation process by reducing the number of unnecessary packages being installed.