chore: Update setup.sh to source config.sh and timeshift_setup.sh
This commit updates the setup.sh script to source the config.sh file and the timeshift_setup.sh script. Previously, the config.sh file was not being sourced, which could cause issues when running the script. Sourcing the config.sh file ensures that the necessary variables are available for use in the script. Additionally, the timeshift_setup.sh script was not being executed, but now it is included in the setup process. This change ensures that the Timeshift setup is performed as intended.

Recent user commits:
- chore: Update timeshift_setup.sh to source functions.sh
- chore: Update setup.sh to remove unnecessary packages from installation
- chore: Update config.sh to export stow_source_dir
- chore: Update setup.sh and stow.sh to print debug information
- chore: Update setup.sh to fix config sourcing path
- chore: Update config sourcing path for setup.sh script